# docs/openapi.yaml
openapi: 3.1.0
info:
  title: GELB CRM API
  version: 0.1.0
  description: API для MVP CRM языковой онлайн-школы
servers:
  - url: /api/v1
security:
  - cookieAuth: []
...
paths:
...
  /teachers:
    get:
      tags: [Teachers]
      summary: Список преподавателей
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Teacher' }
    post:
      tags: [Teachers]
      summary: Создать преподавателя
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TeacherCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Teacher' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: gelb_session
...
  schemas:
...
    Teacher:
      type: object
      properties:
        id: { type: string, format: uuid }
        first_name: { type: string }
        last_name: { type: string }
        full_name: { type: string }
        language_id: { type: integer, nullable: true }
        language_name: { type: string, nullable: true }
        language_flag_emoji: { type: string, nullable: true }
        rate_rub: { type: integer, nullable: true }
        telegram_raw: { type: string, nullable: true }
        telegram_display: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        email: { type: string, format: email, nullable: true }
        comment: { type: string, nullable: true }
        active_students_count: { type: integer }
        created_at: { type: string }
        updated_at: { type: string }
        deleted_at: { type: string, nullable: true }
    TeacherCreate:
      type: object
      required: [firstName, lastName]
      properties:
        firstName: { type: string }
        lastName: { type: string }
        languageId: { type: integer, nullable: true }
        rateRub: { type: integer, nullable: true }
        telegramRaw: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        email: { type: string, format: email, nullable: true }
        comment: { type: string, nullable: true }
...